ragdoll_sleepaftertime on ragdolls_medium (#104)
Changing value here, because there's a chance to play a specific death animation on Medic that is more than 1 second to finish (about 1.2 seconds or below). When this animation plays and the ragdoll_sleepaftertime is with the value of 1, the animation freezes in the mid-air. Just putting 0.5 more here to fix that.