match_hud_alias3
This fixes a rare case when someone uses run_modules but doesn't change classes, so match_hud_alias2 will not be ran again (since it is on game_overrides.cfg). But, with this change, match_hud_off will always be intedeed to disable the match HUD if match_hud_alias2 has been ran before.