Fix match_hud_off not working when run_modules has been used
This fixes a rare case when someone uses run_modules but doesn't change classes, so `match_hud_alias2` will not be run again (since it is on `game_overrides.cfg`). But with this change, `match_hud_off` will always disable the match HUD if `match_hud_alias2` was ran before.