Bolded critical parts of console commands plus grammar fix.
Well the last change was heavily unnecessary but I think this might at least get a bigger visual look to people that glance without it requiring more buttons to be pressed. Plus the current grammar looks like it was written in a rush, or at least to me as a native English speaker it looks wrong. It is missing punctuation, capitalization and feels too shortened.